# docs/swagger.yaml
definitions:
  controller.categoryInput:
    properties:
      jenis_barang:
        type: string
    required:
    - jenis_barang
    type: object
  controller.productInput:
    properties:
      category_id:
        description: JenisBarang      string `json:"jenis_barang" binding:"required"`
        type: integer
      jumlah_terjual:
        type: integer
      nama_barang:
        type: string
      stok:
        type: integer
      tanggal_transaksi:
        type: string
    required:
    - category_id
    - jumlah_terjual
    - nama_barang
    - stok
    - tanggal_transaksi
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jenis_barang:
        type: string
      updated_at:
        type: string
    type: object
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        description: JenisBarang      string    `json:"jenis_barang"`
        type: integer
      created_at:
        type: string
      id:
        type: integer
      jumlah_terjual:
        type: integer
      nama_barang:
        type: string
      stok:
        type: integer
      tanggal_transaksi:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /category:
    get:
      description: Get a list of Category.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get All Category.
      tags:
      - Category
    post:
      description: Add a new Category.
      parameters:
      - description: The body to create a new Category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.categoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Add New Category.
      tags:
      - Category
  /category/{id}:
    delete:
      description: Delete a category by id.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete one category.
      tags:
      - Category
    patch:
      description: Update category by id.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: The body to update an Category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.categoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Update Category.
      tags:
      - Category
  /product:
    get:
      description: Get a list of Product.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get All Product.
      tags:
      - Product
    post:
      description: Add a new Product.
      parameters:
      - description: The body to create a new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.productInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Add New Product.
      tags:
      - Product
  /product/{id}:
    delete:
      description: Delete a product by id.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete one product.
      tags:
      - Product
    patch:
      description: Update product by id.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: The body to update an Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.productInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Update Product.
      tags:
      - Product
swagger: "2.0"